djangobackend/requirements.txt
asgiref==3.8.1
boto3==1.26.165
botocore==1.29.165
Django==4.2
djangorestframework==3.14.0
gunicorn==21.2.0
jmespath==1.0.1
python-dateutil==2.9.0.post0
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.18
